requirements.txt
Django==3.0.8
django-bulk-update-or-create==0.2.0
django-glrm==1.1.3
django-mailgun==0.9.1
django-widget-tweaks==1.4.8
huey==2.3.0
pika==1.1.0
## The following requirements were added by pip freeze:
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
idna==2.10
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
